+++ src/pages/index.astro
@@ -56,7 +56,7 @@ const recentStories = stories
         </a>
         <a 
           href="#projects" 
-          class="px-8 py-4 border border-gray-300 text-gray-700 rounded-lg hover:bg-gray-50 transition-all duration-300 font-semibold"
+          class="px-8 py-4 bg-gradient-to-r from-purple-600/80 to-blue-600/80 backdrop-blur-md text-white rounded-lg hover:shadow-lg transition-all duration-300 transform hover:scale-105 font-semibold"
         >
           프로젝트 보기
         </a>
